+++ src/maddening/examples/basics/bouncing_ball_terminal.py
@@ -9,17 +9,8 @@ Press Ctrl-C to stop.
 
 Usage
 -----
-    cd /home/nick/MSF/MADDENING
-    source ../venvs/.maddening/bin/activate
-    python maddening/examples/bouncing_ball_terminal.py
+    python -m maddening.examples.basics.bouncing_ball_terminal
 """
-
-import sys
-import os
-
-_project_root = os.path.abspath(os.path.join(os.path.dirname(__file__), os.pardir, os.pardir))
-if _project_root not in sys.path:
-    sys.path.insert(0, _project_root)
 
 from maddening.core.graph_manager import GraphManager
 from maddening.nodes.ball import BallNode